salary datatype changed from long to double

+++ practise/src/practisegit/Employee.java
@@ -6,7 +6,7 @@ public class Employee {
 	private String name;
 	private String state;
 	private String address; // Conflicts resolved
-	private long salary;	
+	private double salary;	
 	private int age;
 
 }
